Improve event hierarchy for Asset
