+++ spi/common/core-spi/src/main/java/org/eclipse/edc/spi/event/asset/AssetDeleted.java
@@ -15,19 +15,23 @@
 
 package org.eclipse.edc.spi.event.asset;
 
-import org.eclipse.edc.spi.event.Event;
-
-import java.util.Objects;
-
 /**
  * Describe an Asset deletion, after this has emitted, the Asset represented by the id won't be available anymore.
  */
-public class AssetDeleted extends Event<AssetDeleted.Payload> {
+public class AssetDeleted extends AssetEvent<AssetDeleted.Payload> {
 
     private AssetDeleted() {
     }
 
-    public static class Builder extends Event.Builder<AssetDeleted, Payload, Builder> {
+    /**
+     * This class contains all event specific attributes of a Asset Deletion Event
+     *
+     */
+    public static class Payload extends AssetEvent.Payload {
+
+    }
+
+    public static class Builder extends AssetEvent.Builder<AssetDeleted, Payload, Builder> {
 
         public static Builder newInstance() {
             return new Builder();
@@ -36,23 +40,6 @@ public class AssetDeleted extends Event<AssetDeleted.Payload> {
         private Builder() {
             super(new AssetDeleted(), new Payload());
         }
-
-        public Builder assetId(String assetId) {
-            event.payload.assetId = assetId;
-            return this;
-        }
-
-        @Override
-        protected void validate() {
-            Objects.requireNonNull(event.payload.getAssetId());
-        }
     }
 
-    /**
-     * This class contains all event specific attributes of a Asset Deletion Event
-     *
-     */
-    public static class Payload extends AssetEventPayload {
-
-    }
 }
